wattx/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
django-environ==0.8.1
django-extensions==3.1.5
paho-mqtt==1.6.1
sqlparse==0.4.2
tzdata==2021.5
